Apply code review polish to components and blocks
- Extract category tile illustrations into block-categories-illustrations.tsx
  so block-categories.tsx focuses on the registry and grid logic
- Remove invalid aria-disabled from the non-interactive coming-soon tile
- Announce theme paste status to screen readers (role=status, aria-live)
  and cap the paste textarea at 10k characters
- Name the tree-view indentation constants instead of inline magic numbers
- Document the password input autoComplete override for signup forms
